pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "as-databricks-mlops"
version = "0.2.0"
description = "Lightweight MLOps scaffolding for Databricks projects"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "databricks-sdk>=0.20",
]

[project.scripts]
adm = "as_databricks_mlops.cli:cli"
az-mlops = "az_mlops.cli:cli"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pyyaml>=6.0",
]

[tool.hatch.build.targets.wheel]
packages = ["as_databricks_mlops", "az_mlops"]

[tool.pytest.ini_options]
testpaths = ["tests"]
